fix : 게시글 이미지및 조회 오류 해결,  게시글 삭제 수정

+++ runwithme/src/main/java/com/runwithme/runwithme/domain/challenge/service/ChallengeService.java
@@ -69,8 +69,15 @@ public class ChallengeService {
 	}
 
 	@Transactional
-	public void deleteBoard(Long challengeSeq, Long boardSeq) {
-		challengeRepository.findById(challengeSeq).orElseThrow(() -> new CustomException(USER_NOT_FOUND));
+	public void deleteBoard(Long boardSeq) {
+		final User user = authUtils.getLoginUser();
+
+		final ChallengeBoard board = challengeBoardRepository.findById(boardSeq).orElseThrow(() -> new CustomException(BOARD_NOT_FOUND));
+
+		if (board.getUser().getSeq() != user.getSeq()) {
+			throw new CustomException(BOARD_NOT_CREATE_USER);
+		}
+
 		challengeBoardRepository.deleteById(boardSeq);
 	}
 
@@ -224,6 +231,12 @@ public class ChallengeService {
 			return ImageCache.get(ImageCache.DEFAULT_CHALLENGE);
 		}
 		return imageService.save(image);
+	}
+
+	public Resource getBoardImage(Long boardSeq) {
+		final ChallengeBoard board = challengeBoardRepository.findById(boardSeq)
+			.orElseThrow(() -> new CustomException(BOARD_NOT_FOUND));
+		return imageService.getImage(board.getImage().getSeq());
 	}
 
 	public Resource getChallengeImage(Long challengeSeq) {
